+++ search/bench.go
@@ -62,7 +62,7 @@ var fens = []string{
 
 func RunBenchmark() {
 
-	cacheSize := uint32(32)
+	cacheSize := 32
 	depth := int8(12)
 	runner := NewRunner(NewCache(cacheSize), 1)
 	nodes := int64(0)
